Move get financial accounts query to new folder
