+++ Source/Application/FinancialAccounts/Queries/GetFinancialAccounts/GetFinancialAccountsQuery.cs
@@ -1,10 +1,8 @@
 ﻿using System.Collections.Generic;
-
 using MakeMeRich.Application.Common.Dtos;
-
 using MediatR;
 
-namespace MakeMeRich.Application.FinancialAccounts.Queries
+namespace MakeMeRich.Application.FinancialAccounts.Queries.GetFinancialAccounts
 {
     public class GetFinancialAccountsQuery : IRequest<List<FinancialAccountDto>>
     {
